Move PhysicalBone3D to its own file.
