+++ editor/plugins/skeleton_3d_editor_plugin.cpp
@@ -38,8 +38,8 @@
 #include "node_3d_editor_plugin.h"
 #include "scene/3d/collision_shape_3d.h"
 #include "scene/3d/mesh_instance_3d.h"
+#include "scene/3d/physical_bone_3d.h"
 #include "scene/3d/physics_joint_3d.h"
-#include "scene/3d/rigid_body_3d.h"
 #include "scene/resources/capsule_shape_3d.h"
 #include "scene/resources/sphere_shape_3d.h"
